MAINT: Move the `compare_trajectories` annotations to a dedicated stub file
